backend/requirements.txt
django
djangorestframework
django-cors-headers
cryptography
PyYAML
python-dotenv
paramiko
requests
apscheduler
tftpy
GitPython
boto3
azure-storage-blob
google-cloud-storage
social-auth-app-django
django-auth-ldap
djangosaml2
psycopg2-binary
mysqlclient
python-dateutil